Distance filter update
Added kmeans_filter() function as an alias for when user only want to use k-means clusters to filter out based on pca distance without plotting. This is also now the recommended way to filter, so that the data isn't accidentally filtered multiple times, once per function call. Generally, operations should be done as follows: the filter should be applied first, so that it updates the entire dataset. Afterwards, calling kmeans() to plot or export_cell_lists() to sort cells should be done without passing in a value for distance_filter, as the filter has already been applied.
Changed how the kmeans() function is structured, so that plotting, filtering, and returning the cluster information array can all be done independently.